Move primite declarations and impls into submodule
